Use official Minio Chart instead of custom one
the helmfile.d/0020.minio.yaml mentioned that the minio deployment is
based on the minio operator ... but that's not true. It's a self written
helm chart without using the CRDs of the minio operator!

Our motivation was to replace the custom helm deployment with the
upstream one from minio. So we can enable persistence, without adding
that as a feature to the self written helm chart.

Co-authored-by: Petra Scherer <scherer@synyx.de>
Co-authored-by: Timo Klenk <klenk@synyx.de>
Signed-off-by: Johannes Graf <graf@synyx.de>